+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 ### Added
 
 ### Fixed
+* Fix a possible UI freeze when an output directory is used with more than 10000 subdirectories
+* Fix a few cases which could freeze the UI.
+* Fix exception when running LaTeX on WSL with PyCharm 2025.1+ 
 * Fix a caching issue when building the file set
 * Fix a rare case where an autocompile rerun would try to start before the previous run finished
 * Fix incorrect slashes in path on Windows after pasting image
-* Fix exception #4116
+* Fix exceptions #4116, #4129
 
 ## [0.10.4] - 2025-07-01